==> jupyterLab/Project_1_errorAnalysis/1.densityvenn_P1/0.1_Handle with errPos.ipynb <==
        with open(self.data_file_path, "r") as f:
            header = f.readline().strip().split("\t")
            header.append("COV_Ratio")
            for line in f:
                data.append(line.strip().split("\t") + [None])

def process(self):
    data = []
    missing_ids_count = 0  # Counter for missing IDs
    missing_ids_file = "errPOS_missing_files_COV.txt"  # Filename for missing IDs

    with open(self.data_file_path, "r") as f:
        header = f.readline().strip().split("\t")
        header.append("COV_Ratio")
        for line_count, line in enumerate(f):
            if line_count >= 10:
                break
            data.append(line.strip().split("\t") + [None])

    # Create/Open a txt file to store the IDs of non-existent files
    with open(missing_ids_file, "a") as missing_file:
        for line in data:
            id_now = str(line[0])
            id_position = line[1]
            file_found = False
            if id_position is not None:
                for filename in os.listdir(self.dir_to_search_Cov):
                    if id_now in filename and filename.endswith(".txt"):
                        file_path = os.path.join(self.dir_to_search_Cov, filename)
                        result = self._search_position_value_and_get_column_4(file_path, int(id_position))
                        if result is not None:
                            line[-1] = result
                            file_found = True
                            break
            if not file_found:
                # Write the ID into the txt file
                missing_file.write(str(id_now) + "\t")

                missing_ids_count += 1  # Increment the counter


## 0.For_OUTPUT

import os

file_path = "/nfs/research/goldman/zihao/compViridian_2/MAPLE_Part/For_col/MAPLE0.3.2_rateVar_checkingErrors_col_estimatedErrors.txt"
output_folder = "/nfs/research/goldman/zihao/code/"

# Open the output file to write the processed data
with open(os.path.join(output_folder, "output_modified.txt"), "w") as output_file:
    # Write the column headers to the output file
    output_file.write("ID\tPosition\n")
    
    # Initialize the current_id variable to None
    current_id = None
    
    # Iterate over each line in the file and convert it to a list
    with open(file_path, "r") as input_file:
        for line in input_file:
            if line.startswith(">"):
                # Update current_id if a new ID is encountered
                current_id = line[1:].strip()
            else:
                tokens = line.strip().split()
                position = int(tokens[0])
                base = tokens[1]
                percentage = float(tokens[2])

                # Check if percentage is less than 0.5, and if so, skip adding it to the processed data
                if percentage < 0.5:
                    continue

                # Write the processed data to the output file
                output_file.write(f"{current_id}\t{position}\n")
                
                # Delete variables to free up memory
                del tokens, base, position, percentage
                
    # Delete the current_id variable after the loop has finished
    del current_id

print(f"Processing complete and file written to {output_folder}.")

file_path = '/nfs/research/goldman/zihao/code/output_modified.txt'

with open(file_path, 'r') as file:
    line_count = sum(1 for line in file)

print(f"The number of MAPLE-marked error positions is: {line_count-1}")

## 1.For_COV

import os
import numpy as np

def search_position_value_and_get_column_4(file_path, id_position):
    with open(file_path, "r") as f:
        file_header = f.readline().strip().split("\t")
        position_column_index = file_header.index("Position")
        column_4_index = 2

        for line in f:
            line_values = line.strip().split("\t")
            if position_column_index < len(line_values): # Check if index is within range
                position_value = int(line_values[position_column_index])
                if position_value == id_position and column_4_index < len(line_values):
                    return line_values[column_4_index]
    return None

def calculate_mean_error(data):
    position_error_sum = {}
    position_count = {}

    for line in data:
        position = line[1]
        mean_error = line[-1]

        if mean_error != "None":
            mean_error = float(mean_error)

            if position in position_error_sum:
                position_error_sum[position] += mean_error
                position_count[position] += 1
            else:
                position_error_sum[position] = mean_error
                position_count[position] = 1

    mean_error_by_position = {}
    for position in position_error_sum:
        mean_error_by_position[position] = position_error_sum[position] / position_count[position]

    return mean_error_by_position

def process_data(data_file_path, directory_to_search, output_file_path):
    data = []

    with open(data_file_path, "r") as f:
        header = f.readline().strip().split("\t")
        header.append("MEAN_err")

        for line in f:
            data.append(line.strip().split("\t") + [None])

    for line in data:
        id_now = str(line[0])
        id_position = line[1]

        if id_position is not None:
            for filename in os.listdir(directory_to_search):
                if id_now in filename and filename.endswith(".txt"):
                    file_path = os.path.join(directory_to_search, filename)
                    result = search_position_value_and_get_column_4(file_path, int(id_position))
                    if result is not None:
                        line[-1] = result
                        break

    # Remove rows with "None" or None in the MEAN_err column
    data = [line for line in data if line[-1] is not None and line[-1] != "None"]

    # Calculate the mean error for each position
    mean_error_by_position = calculate_mean_error(data)

    # Write the result to the output file
    with open(output_file_path, "w") as f:
        f.write("Position\tMEAN_err\n")
        for position, mean_error in mean_error_by_position.items():
            f.write(f"{position}\t{mean_error}\n")
            
directory_to_search = "/nfs/research/goldman/zihao/Datas/p1/File_5_coverage/Decompress/"
data_file_path = os.path.join(output_folder, "output_modified.txt")
output_file_path = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing_forCoverage/COV_dataProcessing_errorPos.txt"

process_data(data_file_path, directory_to_search, output_file_path)

## 2.For_ANNOT

import os

class DataProcessor:
    def __init__(self, data1_file, data_folder, output_file):
        self.data1_file = data1_file
        self.data_folder = data_folder
        self.output_file = output_file

    def read_data1(self):
        """Read data1 from file and return the header and data."""
        data1 = []
        with open(self.data1_file) as f:
            header = f.readline().strip() + "\tAF\tSB\n"
            for line in f:
                line = line.strip().split()
                data1.append(line)
        return header, data1

    def process_data(self):
        """Process data1 by looking up corresponding rows in data folder."""
        header, data1 = self.read_data1()

        for row in data1:
            id = row[0]
            filename = os.path.join(self.data_folder, f"{id}_annot.txt")
            if os.path.exists(filename):
                with open(filename) as f:
                    for line in f:
                        line = line.strip().split()
                        if line[0] == row[1]:
                            row.append(line[3])
                            row.append(line[4])
                            break
                    else:
                        row.append("")
                        row.append("")
            else:
                row.append("")
                row.append("")

        return header, data1

    def write_output(self):
        """Write the processed data1 to the output file."""
        header, data1 = self.process_data()

        with open(self.output_file, "w") as f:
            f.write(header)
            for row in data1:
                f.write("\t".join(row) + "\n")
if __name__ == '__main__':
    data1_file = os.path.join(output_folder, "output_modified.txt")
    data_folder = "/nfs/research/goldman/zihao/Datas/p1/File_5_annot/New_Decompress/"
    output_file = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing_forCoverage/ANNOT_dataProcessing_errorPos.txt"

    processor = DataProcessor(data1_file, data_folder, output_file)
    processor.write_output()

### [!!!]Combination

###### Run the program

```bash
bsub -M 20000 -e /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/errorChecking_errPos_dataProcessing.txt -o /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/outputChecking_errPos_dataProcessing.txt 'python3 /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/CODE_errPos_dataProcessing.py'
```

import os

class MapleFileProcessor:
    def __init__(self, input_file, output_file):
        self.input_file = input_file
        self.output_file = output_file

    def process(self):
        with open(self.output_file, "w") as output_file:
            output_file.write("ID\tPosition\n")
            current_id = None
            with open(self.input_file, "r") as input_file:
                for line in input_file:
                    if line.startswith(">"):
                        current_id = line[1:].strip()
                    else:
                        tokens = line.strip().split()
                        position = int(tokens[0])
                        percentage = float(tokens[2])
                        if percentage < 0.5:
                            continue
                        output_file.write(f"{current_id}\t{position}\n")


class CovDataProcessor:
    def __init__(self, data_file_path, dir_to_search_Cov, output_file_path):
        self.data_file_path = data_file_path
        self.dir_to_search_Cov = dir_to_search_Cov
        self.output_file_path = output_file_path

    @staticmethod
    def _search_position_value_and_get_column_4(file_path, id_position):
        with open(file_path, "r") as f:
            file_header = f.readline().strip().split("\t")
            position_column_index = file_header.index("Position")
            column_4_index = 2
            for line in f:
                line_values = line.strip().split("\t")
                if position_column_index < len(line_values):
                    position_value = int(line_values[position_column_index])
                    if position_value == id_position and column_4_index < len(line_values):
                        return line_values[column_4_index]
        return None

    @staticmethod
    def _calculate_mean_error(data):
        position_error_sum = {}
        position_count = {}
        for line in data:
            position = line[1]
            mean_error = line[-1]
            if mean_error != "None":
                mean_error = float(mean_error)
                if position in position_error_sum:
                    position_error_sum[position] += mean_error
                    position_count[position] += 1
                else:
                    position_error_sum[position] = mean_error
                    position_count[position] = 1
        mean_error_by_position = {}
        for position in position_error_sum:
            mean_error_by_position[position] = position_error_sum[position] / position_count[position]
        return mean_error_by_position

    def process(self):
        data = []
        with open(self.data_file_path, "r") as f:
            header = f.readline().strip().split("\t")
            header.append("COV_Ratio")
            for line in f:
                data.append(line.strip().split("\t") + [None])
        for line in data:
            id_now = str(line[0])
            id_position = line[1]
            if id_position is not None:
                for filename in os.listdir(self.dir_to_search_Cov):
                    if id_now in filename and filename.endswith(".txt"):
                        file_path = os.path.join(self.dir_to_search_Cov, filename)
                        result = self._search_position_value_and_get_column_4(file_path, int(id_position))
                        if result is not None:
                            line[-1] = result
                            break
        data = [line for line in data if line[-1] is not None and line[-1] != "None"]
        mean_error_by_position = self._calculate_mean_error(data)
        with open(self.output_file_path, "w") as f:
            f.write("Position\tCOV_Ratio\n")
            for position, mean_error in mean_error_by_position.items():
                f.write(f"{position}\t{mean_error}\n")


class AnnotDataProcessor:
    def __init__(self, data1_file, dir_to_search_Annot, output_file):
        self.data1_file = data1_file
        self.dir_to_search_Annot = dir_to_search_Annot
        self.output_file = output_file

    def _read_data1(self):
        data1 = []
        with open(self.data1_file) as f:
            header = f.readline().strip() + "\tAF\tSB\n"
            for line in f:
                line = line.strip().split()
                data1.append(line)
        return header, data1

    def _process_data(self):
        header, data1 = self._read_data1()
        for row in data1:
            id = row[0]
            filename = os.path.join(self.dir_to_search_Annot, f"{id}_annot.txt")
            if os.path.exists(filename):
                with open(filename) as f:
                    for line in f:
                        line = line.strip().split()
                        if line[0] == row[1]:
                            row.append(line[3])
                            row.append(line[4])
                            break
                    else:
                        row.append("")
                        row.append("")
            else:
                row.append("")
                row.append("")
        return header, data1

    def process(self):
        header, data1 = self._process_data()
        with open(self.output_file, "w") as f:
            f.write(header)
            for row in data1:
                f.write("\t".join(row) + "\n")


if __name__ == '__main__':
    input_file = "/nfs/research/goldman/zihao/compViridian_2/MAPLE_Part/For_col/MAPLE0.3.2_rateVar_checkingErrors_col_estimatedErrors.txt"
    output_folder = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/"
    output_file = os.path.join(output_folder, "output_modified.txt")
    mapleProcessor = MapleFileProcessor(input_file, output_file)
    mapleProcessor.process()

    dir_to_search_Cov = "/nfs/research/goldman/zihao/Datas/p1/File_5_coverage/Decompress/"
    data_file_path = output_file
    output_file_cov = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/COV_dataProcessing_errorPos.txt"
    covProcessor = CovDataProcessor(data_file_path, dir_to_search_Cov, output_file_cov)
    covProcessor.process()

    data1_file = output_file
    dir_to_search_Annot = "/nfs/research/goldman/zihao/Datas/p1/File_5_annot/New_Decompress/"
    output_file_annot = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/ANNOT_dataProcessing_errorPos.txt"
    annotProcessor = AnnotDataProcessor(data1_file, dir_to_search_Annot, output_file_annot)
    annotProcessor.process()
    

with open(output_file, 'r') as file:
    line_count = sum(1 for line in file)

print(f"The number of MAPLE-marked error positions is: {line_count-1}")

with open(output_file_cov, 'r') as file:
    line_count = sum(1 for line in file)
print(f"The number of MAPLE-marked error positions "
      f"(After finishing the COV information merge operation) is: {line_count-1}")

with open(output_file_annot, 'r') as file:
    line_count = sum(1 for line in file)
print(f"The number of MAPLE-marked error positions "
      f"(After finishing the ANNOT information merge operation) is: {line_count-1}")

###### For output

```bash
bsub -M 2000 -e /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/Folder_Checking/errorChecking_errPos_dataProcessing_forOutput.txt -o /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/Folder_Checking/outputChecking_errPos_dataProcessing_forOutput.txt 'python3 /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/CODE_errPos_dataProcessing_forOutput.py'
```

import os

class MapleFileProcessor:
    def __init__(self, input_file, output_file):
        self.input_file = input_file
        self.output_file = output_file

    def process(self):
        with open(self.output_file, "w") as output_file:
            output_file.write("ID\tPosition\n")
            current_id = None
            with open(self.input_file, "r") as input_file:
                for line in input_file:
                    if line.startswith(">"):
                        current_id = line[1:].strip()
                    else:
                        tokens = line.strip().split()
                        position = int(tokens[0])
                        percentage = float(tokens[2])
                        if percentage < 0.5:
                            continue
                        output_file.write(f"{current_id}\t{position}\n")

if __name__ == '__main__':
    input_file = "/nfs/research/goldman/zihao/compViridian_2/MAPLE_Part/For_col/MAPLE0.3.2_rateVar_checkingErrors_col_estimatedErrors.txt"
    output_folder = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/"
    output_file = os.path.join(output_folder, "output_modified.txt")
    mapleProcessor = MapleFileProcessor(input_file, output_file)
    mapleProcessor.process()

with open(output_file, 'r') as file:
    line_count = sum(1 for line in file)

print(f"The number of MAPLE-marked error positions is: {line_count-1}")

###### For cov

```bash
bsub -M 2000 -e /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/Folder_Checking/errorChecking_errPos_dataProcessing_forCov.txt -o /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/Folder_Checking/outputChecking_errPos_dataProcessing_forCov.txt 'python3 /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/CODE_errPos_dataProcessing_forCov.py'
```

import os

class CovDataProcessor:
    def __init__(self, data_file_path, dir_to_search_Cov, output_file_path):
        self.data_file_path = data_file_path
        self.dir_to_search_Cov = dir_to_search_Cov
        self.output_file_path = output_file_path

    @staticmethod
    def _search_position_value_and_get_column_4(file_path, id_position):
        with open(file_path, "r") as f:
            file_header = f.readline().strip().split("\t")
            position_column_index = file_header.index("Position")
            column_4_index = 2
            for line in f:
                line_values = line.strip().split("\t")
                if position_column_index < len(line_values):
                    position_value = int(line_values[position_column_index])
                    if position_value == id_position and column_4_index < len(line_values):
                        return line_values[column_4_index]
        return None

    @staticmethod
    def _calculate_mean_error(data):
        position_error_sum = {}
        position_count = {}
        for line in data:
            position = line[1]
            mean_error = line[-1]
            if mean_error != "None":
                mean_error = float(mean_error)
                if position in position_error_sum:
                    position_error_sum[position] += mean_error
                    position_count[position] += 1
                else:
                    position_error_sum[position] = mean_error
                    position_count[position] = 1
        mean_error_by_position = {}
        for position in position_error_sum:
            mean_error_by_position[position] = position_error_sum[position] / position_count[position]
        return mean_error_by_position

    def process(self):
        data = []
        with open(self.data_file_path, "r") as f:
            header = f.readline().strip().split("\t")
            header.append("COV_Ratio")
            for line in f:
                data.append(line.strip().split("\t") + [None])
        for line in data:
            id_now = str(line[0])
            id_position = line[1]
            if id_position is not None:
                for filename in os.listdir(self.dir_to_search_Cov):
                    if id_now in filename and filename.endswith(".txt"):
                        file_path = os.path.join(self.dir_to_search_Cov, filename)
                        result = self._search_position_value_and_get_column_4(file_path, int(id_position))
                        if result is not None:
                            line[-1] = result
                            break
        data = [line for line in data if line[-1] is not None and line[-1] != "None"]
        mean_error_by_position = self._calculate_mean_error(data)
        with open(self.output_file_path, "w") as f:
            f.write("Position\tCOV_Ratio\n")
            for position, mean_error in mean_error_by_position.items():
                f.write(f"{position}\t{mean_error}\n")

if __name__ == '__main__':
    output_folder = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/"
    output_file = os.path.join(output_folder, "output_modified.txt")

    
    # Wait until the output_file exists
    while not os.path.exists(output_file):
        time.sleep(5)  # Wait for 5 seconds before checking again
    print("Output file found. Proceeding with the process...")
    
    dir_to_search_Cov = "/nfs/research/goldman/zihao/Datas/p1/File_5_coverage/Decompress/"
    data_file_path = output_file
    output_file_cov = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/COV_dataProcessing_errorPos.txt"
    covProcessor = CovDataProcessor(data_file_path, dir_to_search_Cov, output_file_cov)
    covProcessor.process()

with open(output_file_cov, 'r') as file:
    line_count = sum(1 for line in file)
print('=========================== information =========================== ')
print(f"The number of MAPLE-marked error positions "
      f"(After finishing the COV information merge operation) is: {line_count-1}")

###### For annot

```bash
bsub -M 2000 -e /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/Folder_Checking/errorChecking_errPos_dataProcessing_forAnnot.txt -o /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/Folder_Checking/outputChecking_errPos_dataProcessing_forAnnot.txt 'python3 /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/CODE_errPos_dataProcessing_forAnnot.py'
```

import os

class AnnotDataProcessor:
    def __init__(self, data1_file, dir_to_search_Annot, output_file):
        self.data1_file = data1_file
        self.dir_to_search_Annot = dir_to_search_Annot
        self.output_file = output_file

    def _read_data1(self):
        data1 = []
        with open(self.data1_file) as f:
            header = f.readline().strip() + "\tAF\tSB\n"
            for line in f:
                line = line.strip().split()
                data1.append(line)
        return header, data1

    def _process_data(self):
        header, data1 = self._read_data1()
        for row in data1:
            id = row[0]
            filename = os.path.join(self.dir_to_search_Annot, f"{id}_annot.txt")
            if os.path.exists(filename):
                with open(filename) as f:
                    for line in f:
                        line = line.strip().split()
                        if line[0] == row[1]:
                            row.append(line[3])
                            row.append(line[4])
                            break
                    else:
                        row.append("")
                        row.append("")
            else:
                row.append("")
                row.append("")
        return header, data1

    def process(self):
        header, data1 = self._process_data()
        with open(self.output_file, "w") as f:
            f.write(header)
            for row in data1:
                f.write("\t".join(row) + "\n")


if __name__ == '__main__':
    output_folder = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/"
    output_file = os.path.join(output_folder, "output_modified.txt")

    # Wait until the output_file exists
    while not os.path.exists(output_file):
        time.sleep(5)  # Wait for 5 seconds before checking again
    print("Output file found. Proceeding with the process...")
    
    data1_file = output_file
    dir_to_search_Annot = "/nfs/research/goldman/zihao/Datas/p1/File_5_annot/New_Decompress/"
    output_file_annot = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/ANNOT_dataProcessing_errorPos.txt"
    annotProcessor = AnnotDataProcessor(data1_file, dir_to_search_Annot, output_file_annot)
    annotProcessor.process()
    
with open(output_file_annot, 'r') as file:
    line_count = sum(1 for line in file)
print('=========================== information =========================== ')
print(f"The number of MAPLE-marked error positions "
      f"(After finishing the ANNOT information merge operation) is: {line_count-1}")

```bash
sh /nfs/research/goldman/zihao/N_errorsProject_1/Folder_errPos_dataProcessing/BASH_errPos_dataProcessing.sh
```
