==> jupyterLab/Project_1_errorAnalysis/1.densityvenn_P1/2.1_prepareData_venn_errPOS.ipynb <==
## <center>Preparing data for the Venn diagram.</center>


| **Label** | **start time** | **finish time** | **last modified** |
|:--------------:|:-----------:|:-----------:|:----------------:|
|   Project 1   |  2023-07-10 |  2023-07-19 |   2023-07-19     |

####  Threshold:
- SB > 0, 
- coverage ratio < 0.1,
- AF > 0.001.

import pandas as pd

threshold_AF = 0.001
threshold_SB = 0
threshold_COV = 0.1

## 1. Error position

input_for_annot_err = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/ANNOT_dataProcessing_errorPos.txt"
input_for_cov_err = "/nfs/research/goldman/zihao/Datas/p1_errorsProject_NEW/Folder_dataProcessing/COV_dataProcessing_errorPos.txt"

### For annot
df_err_annot = pd.read_csv(input_for_annot_err, sep='\t')

df_err_annot['ID_POS'] = df_err_annot['ID'].astype(str) + "_" + df_err_annot['Position'].astype(str)

df_err_annot['Flag_AF'] = df_err_annot['AF'] > threshold_AF
df_err_annot['Flag_SB'] = df_err_annot['SB'] > threshold_SB
df_err_annot.drop(['ID', 'Position','AF','SB'], axis=1, inplace=True)

print(df_err_annot)
print('=====================================')
### For coverage
df_err_cov = pd.read_csv(input_for_cov_err, sep='\t')

df_err_cov['ID_POS'] = df_err_cov['ID'].astype(str) + "_" + df_err_cov['Position'].astype(str)

df_err_cov['Flag_COV'] = df_err_cov['COV_Ratio'] < threshold_COV
df_err_cov.drop(['ID', 'Position', 'COV_Ratio'], axis=1, inplace=True)

print(df_err_cov)
print('=====================================')
### Merge
merged_df = pd.merge(df_err_annot, df_err_cov, on='ID_POS', how='inner')
merged_df.to_csv("/nfs/research/goldman/zihao/Code/jupyterLab/A_Datas/ERR_POS_all.txt", sep='\t', index=False)

merged_df

### Next 
- go ---> 
```python
/nfs/research/goldman/zihao/code/EBI_INTER/Project_1_errorAnalysis/1.densityvenn_P1/1.2_vennChart.Rmd
```
