==> jupyterLab/Project_1_errorAnalysis/1.densityvenn_P1/2.2_calculatePercentage_venn_allPOS.ipynb <==
## <center>Preparing data for the Venn diagram.</center>


| **Label** | **start time** | **finish time** | **last modified** |
|:--------------:|:-----------:|:-----------:|:----------------:|
|   Project 1   |  2023-07-10 |  2023-07-19 |   2023-07-19     |

####  Threshold:
- SB > 0, 
- coverage ratio < 0.1,
- AF > 0.001.

====================== Info ======================  
Total number of positions:  4033264429  
nb_SB_PASSED:  1782985  
nb_AF_PASSED:  7547094  
nb_COV_PASSED:  583535685  
====================== Info ======================  
both_PASSED:  377289  
both_UNPASSED:  3420841838  
====================== Info ======================  
nb_AF_SB_PASSED:  17971522  
nb_SB_COV_PASSED:  253433  
nb_AF_COV_PASSED:  954583  

same_count = 4033264429
Flag_SB_PASSED = 1782985
Flag_AF_PASSED = 7547094
Flag_COV_PASSED = 583535685
both_PASSED = 377289
both_UNPASSED = 3420841838
Flag_SB_Flag_AF_PASSED = 17971522
Flag_SB_Flag_COV_PASSED = 253433
Flag_AF_Flag_COV_PASSED = 954583

pass_1 = Flag_SB_PASSED+Flag_AF_PASSED+Flag_COV_PASSED
pass_2 = Flag_SB_Flag_AF_PASSED+Flag_SB_Flag_COV_PASSED+Flag_AF_Flag_COV_PASSED
pass_3 = both_PASSED + both_UNPASSED
nb_pass = pass_1+pass_2+pass_3

print(round(both_UNPASSED/nb_pass*100,3),'%')
print('=======================================')

print(round(Flag_AF_PASSED/nb_pass*100,3),'%')
print(round(Flag_COV_PASSED/nb_pass*100,3),'%')
print(round(Flag_SB_PASSED/nb_pass*100,3),'%')
print('=======================================')

print(round(Flag_SB_Flag_AF_PASSED/nb_pass*100,3),'%')
print(round(Flag_SB_Flag_COV_PASSED/nb_pass*100,3),'%')
print(round(Flag_AF_Flag_COV_PASSED/nb_pass*100,3),'%')

print('=======================================')
print(round(both_PASSED/nb_pass*100,3),'%')
